Add files via upload
# To-Do List Web Application

This web-based to-do list application is a simple and effective way to manage your tasks online. Built with HTML, CSS, and JavaScript, it provides a responsive and intuitive user interface, making task management a breeze for users of all levels.

## Features
- *Dynamic Task Handling:* Add, edit, and remove tasks dynamically without page reloads.
- *Responsive Design:* Ensures a seamless experience across various devices.
- *Local Storage:* Persists tasks locally, allowing users to access their to-do list even after closing the browser.
- *Priority and Due Dates:* Assign priorities and due dates to tasks for better organization.

Customization Options : Easily add and remove tasks with a few clicks.Prioritize tasks using different levels and set due dates for time-sensitive activities.

This to-do list web application is designed to simplify task management, providing users with a visually appealing and interactive platform. Whether you're at your desk or on the go, this responsive application ensures a consistent experience, allowing you to stay organized wherever you are.